add createSession utility
