+++ src/utils/createSession.js
@@ -1,0 +1,14 @@
+import { FIFTEEN_MINUTES, THIRTY_DAYS } from '../constants/index.js';
+import { randomBytes } from 'crypto';
+
+export const createSession = () => {
+    const accessToken = randomBytes(30).toString('base64');
+    const refreshToken = randomBytes(30).toString('base64');
+
+    return {
+        accessToken,
+        refreshToken,
+        accessTokenValidUntil: new Date(Date.now() + FIFTEEN_MINUTES),
+        refreshTokenValidUntil: new Date(Date.now() + THIRTY_DAYS),
+    };
+};
